+++ src/Action/RequestAction.php
@@ -39,7 +39,7 @@ final class RequestAction
         private UserManagerInterface $userManager,
         private MailerInterface $mailer,
         private TokenGeneratorInterface $tokenGenerator,
-        private int $retryTtl
+        private int $retryTtl,
     ) {
     }
 
@@ -66,9 +66,7 @@ final class RequestAction
                 $this->userManager->save($user);
             }
 
-            return new RedirectResponse($this->urlGenerator->generate('sonata_user_admin_resetting_check_email', [
-                'username' => $username,
-            ]));
+            return new RedirectResponse($this->urlGenerator->generate('sonata_user_admin_resetting_check_email'));
         }
 
         return new Response($this->twig->render('@SonataUser/Admin/Security/Resetting/request.html.twig', [