+++ src/Command/ActivateUserCommand.php
@@ -53,14 +53,14 @@ final class ActivateUserCommand extends Command
         $user = $this->userManager->findUserByUsername($username);
 
         if (null === $user) {
-            throw new \InvalidArgumentException(sprintf('User identified by "%s" username does not exist.', $username));
+            throw new \InvalidArgumentException(\sprintf('User identified by "%s" username does not exist.', $username));
         }
 
         $user->setEnabled(true);
 
         $this->userManager->save($user);
 
-        $output->writeln(sprintf('User "%s" has been activated.', $username));
+        $output->writeln(\sprintf('User "%s" has been activated.', $username));
 
         return 0;
     }